pyproject.toml
[project]
name = "benchmark_AD"
version = "0.1.0"
description = "Thesis pipeline: real-time visual defect detection with robustness evaluation"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "numpy",
  "pyyaml",
  "opencv-python",
  "scikit-learn",
  "scipy",
  "pillow",
  "torch",
  "torchvision",
  "anomalib>=2.2,<3.0",
  "lightning>=2.4,<3.0",
  "transformers>=4.44,<5.0",
  "psutil",
  "plotly",
  "pandas",
  "tqdm",
  "umap-learn",
]

[project.optional-dependencies]
gpu = [
  "torch",
  "torchvision",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests", "src/corruptions"]

# ---------------------------------------------------------------------------
# uv-specific: route torch/torchvision to the PyTorch CUDA 12.4 wheel index
# only when the "gpu" extra is requested (`uv sync --extra gpu`).
# CPU installs (`uv sync` / `pip install -e .`) pull from PyPI as normal.
# ---------------------------------------------------------------------------

[[tool.uv.index]]
name = "pytorch-cu124"
url  = "https://download.pytorch.org/whl/cu124"
explicit = true   # only used when a [tool.uv.sources] rule references it

[tool.uv.sources]
torch       = [
  { index = "pytorch-cu124", extra = "gpu" },
]
torchvision = [
  { index = "pytorch-cu124", extra = "gpu" },
]
